pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "climb-sync"
version = "0.2.0"
description = "Sync Wahoo Kickr Climb tilt to Zwift road grade during ERG workouts"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "coffeefueled", email = "coffeefueled@gmail.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "websockets>=15.0",
    "zeroconf>=0.148",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
app = [
    "pystray==0.19.5",
    "Pillow>=10.0",
]
build = [
    "pyinstaller==6.19.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/climb_sync"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
